+build ignore willf/bloom benchmarks
Their appearance in a non-ignored file made this package and all its
dependencies a transitive dependency for us. If I understand Go modules
correctly, only the go.mod for such a dependency needs to be fetched,
but it still looks ugly.

The go.sum still contains the packages for reproducibility.