enhance: Expose Locations() and TestLocations API
When dealing with a collection of Bloom filters, determining which filter will return true for a
given key can be computationally intensive, as it requires checking the key's locations in each Bloom filter.

To address this, similar to the implementation in https://github.com/bits-and-blooms/bloom,
it would be beneficial to expose the Locations() and TestLocations() APIs.
This way, we can reuse the locations of a given key, improving efficiency.